// src/generated/ChatRuntimeCommandErrorCode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ChatRuntimeCommandErrorCode = "emptyMessage" | "chatNotFound" | "setupIncomplete" | "notActive" | "runtimeUnavailable" | "authenticationRequired" | "conversationFailed" | "storageUnavailable";