// src/generated/ConversationPromptRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConversationStreamingBehavior } from "./ConversationStreamingBehavior";

export type ConversationPromptRequest = { chatId: string, streamingBehavior: ConversationStreamingBehavior, text: string, };